Support Journal files created by Fleet Carrier Update game versions
At least the events that failed to parse correctly from my journal files from
2020/2021. I added all failing log events (out of ~160k) to relevant test
files.

+++ src/modules/logs/content/log_event_content/fsd_target_event.rs
@@ -11,6 +11,6 @@ pub struct FSDTargetEvent {
     name: String,
 
     #[serde(default)]
-    remaining_jumps_in_route: u8,
+    remaining_jumps_in_route: u16,
     star_class: StarClass,
 }
